Revert "fix: add int32/int type conversions for puregotk app.Run API"
This reverts commit ba73182a3dcc4d562988295b9a78a6d815ddcb3c.

+++ pkg/adwutil/examples/basic/main.go
@@ -24,8 +24,8 @@ func main() {
 	}
 	app.ConnectActivate(&onActivate)
 
-	if code := app.Run(int32(len(os.Args)), os.Args); code > 0 {
-		os.Exit(int(code))
+	if code := app.Run(len(os.Args), os.Args); code > 0 {
+		os.Exit(code)
 	}
 }
